UI: Refactor MockContextProvider
