+++ HoloRepositoryUI/.dockerignore
@@ -2,6 +2,7 @@
 **/node_modules/
 **/dist/
 **/build/
+**/__test_utils__/
 
 # Note: Can be just all in **/__tests__/ after JSONs in test directory not needed for static mock data anymore
 **/__tests__/**/*.ts
